sellshop/requirements.txt
altgraph==0.17.2
amqp==5.1.1
asgiref==3.5.0
asttokens==2.0.5
async-timeout==4.0.2
auto-py-to-exe==2.18.0
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
bottle==0.12.19
bottle-websocket==0.2.9
cached-property==1.5.2
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.2
django-4-jet==1.0.8
django-admin-easy==0.6.1
django-adminlte-3==0.1.6
django-baton==2.3.0
django-celery-beat==2.3.0
django-ckeditor==6.2.0
django-cors-headers==3.13.0
django-countries==7.3.2
django-extensions==3.1.5
django-feedparser==0.2.1
django-filter==21.1
django-jazzmin==2.5.0
django-jet==1.0.8
django-js-asset==2.0.0
django-modeltranslation==0.17.7
django-nested-admin==3.4.0
django-phonenumber-field==6.3.0
django-phonenumbers==1.0.1
django-timezone-field==5.0
django-url-filter==0.3.15
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
Eel==0.12.4
enum-compat==0.0.3
environs==9.5.0
executing==0.8.3
feedparser==6.0.10
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
inflection==0.5.1
ipython==8.4.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
kombu==5.2.4
Markdown==3.3.7
MarkupSafe==2.1.1
marshmallow==3.16.0
matplotlib-inline==0.1.3
oauthlib==3.2.0
packaging==21.3
parso==0.8.3
pexpect==4.8.0
phonenumbers==8.12.51
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
pyinstaller==4.9
pyinstaller-hooks-contrib==2022.2
PyJWT==2.3.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-monkey-business==1.0.0
python-slugify==6.1.2
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2022.1
redis==4.3.3
redis-tools==0.0.2
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sgmllib3k==1.0.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
stack-data==0.3.0
text-unidecode==1.3
traitlets==5.3.0
typing-extensions==4.1.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
webcolors==1.11.1
whichcraft==0.6.1
wrapt==1.14.1
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
